refactor project hook
