+++ apps/web/client/src/components/store/project/manager.ts
@@ -3,11 +3,11 @@ import { makeAutoObservable } from 'mobx';
 
 // Stubs for now
 export class DomainsManager {
-    constructor() {}
+    constructor() { }
 }
 
 export class VersionsManager {
-    constructor(private projectManager: ProjectManager) {}
+    constructor(private projectManager: ProjectManager) { }
 }
 
 export class ProjectManager {
@@ -41,5 +41,5 @@ export class ProjectManager {
         this.project = newProject;
     }
 
-    dispose() {}
+    dispose() { }
 }
