+++ apps/web/client/src/components/store/projects/index.ts
@@ -1,6 +1,7 @@
 import { createContext, useContext } from 'react';
+import { userManager } from '../user';
 import { ProjectsManager } from './manager';
 
-const projectsManager = new ProjectsManager();
+const projectsManager = new ProjectsManager(userManager);
 const ProjectsContext = createContext(projectsManager);
 export const useProjectsManager = () => useContext(ProjectsContext);
